backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
python-docx
spacy
nltk
scikit-learn
google-generativeai
reportlab
python-dotenv
pydantic